Legger hentFagsakForPerson inn i useFagsakApi, fjerner feilhåndtering som ikke gjorde noe (da er det bedre med en TODO)
