+++ src/frontend/sider/Oppgavebenk/OppgaveDirektelenke.tsx
@@ -6,11 +6,11 @@ import { Button, Loader } from '@navikt/ds-react';
 import { RessursStatus } from '@navikt/familie-typer';
 
 import { useApp } from '../../context/AppContext';
-import { useFagsakContext } from '../../context/fagsak/FagsakContext';
 import { AlertType, ToastTyper } from '../../komponenter/Toast/typer';
 import type { IOppgave } from '../../typer/oppgave';
 import { oppgaveTypeFilter, OppgavetypeFilter } from '../../typer/oppgave';
 import { hentFnrFraOppgaveIdenter } from '../../utils/oppgave';
+import useFagsakApi from '../Fagsak/useFagsakApi';
 
 interface IOppgaveDirektelenke {
     oppgave: IOppgave;
@@ -18,7 +18,7 @@ interface IOppgaveDirektelenke {
 
 const OppgaveDirektelenke: React.FC<IOppgaveDirektelenke> = ({ oppgave }) => {
     const { settToast } = useApp();
-    const { hentFagsakForPerson } = useFagsakContext();
+    const { hentFagsakForPerson } = useFagsakApi();
     const { sjekkTilgang } = useApp();
     const [laster, settLaster] = useState<boolean>(false);
     const navigate = useNavigate();
